+++ Backend/main/Account MS/Account-MS/src/main/java/csd/backend/Account/MS/Repository/Player/PlayerChampionStatsRepository.java
@@ -6,6 +6,6 @@ import csd.backend.Account.MS.Model.Player.*;
 import java.util.*;
 
 public interface PlayerChampionStatsRepository extends JpaRepository<PlayerChampionStats, Long>{
-    PlayerChampionStats findByPlayerIdAndChampionId(Long playerId, int championId);
+    PlayerChampionStats findByPlayerIdAndChampionId(Long playerId, Long championId);
     List<PlayerChampionStats> findByPlayerId(Long playerId);
 }
